// src/resources/types/FlagTypes.ts
export enum Flag {
    FIRST_STRIKE = 0,
    CHARGEABLE = 1,
    IGNITES = 2,
    NEGATE_FIRE_DAMAGE = 3,
    NEGATE_BLOCK_OVERFLOW_MODIFIER = 4,
    HAS_BLOCK_MODIFER = 5,
    MODIFIER_ROLE = 6,
    MODIFIER_TYPE = 7,
    MODIFIER_AMOUNT = 8,
    MODIFIER_REMOVE_AFTER_USE = 9,
    READ_OPPONENT = 10,
    RESELECT_SPELLS = 11,
    GAIN_MANA_FROM_DAMAGE = 12,
    ADD_MODIFIER = 13, //NEEDED ? 
    SELF_INFLICTED_DAMAGE = 14,
    FREEZES = 15,
    BLOCK_MODIFIER_TYPE = 16,
}